rewrite logic

+++ src/file_service/file_service.py
@@ -1,14 +1,21 @@
 import os
+from src import utils
 
 
 def read(filename):
-    with open(filename, 'r') as file:
-        data = file.read()
-        return data
+    if os.path.isfile(filename) and (not os.path.isdir(filename)):
+        with open(filename, 'r') as file:
+            return file.read()
+    else:
+        return False
 
 
 def delete(filename):
-    os.remove(filename)
+    if os.path.isfile(filename):
+        os.remove(filename)
+        return True
+    else:
+        return False
 
 
 def list_dir():
@@ -16,17 +23,32 @@ def list_dir():
 
 
 def change_dir(directory):
-    os.chdir(directory)
+    if os.path.isdir(directory):
+        os.chdir(directory)
+        return True
+    else:
+        return False
 
 
-def create(filename, content):
+def create(content):
+    filename = utils.generate_name(10)
+    create_file(filename, content)
+    return filename
+
+
+def create_file(filename, content):
     with open(filename, "w") as file:
         file.write(content)
 
 
 def get_permissions(filename):
-    return oct(os.stat(filename).st_mode)
+    if os.path.isfile(filename) and (not os.path.isdir(filename)):
+        return oct(os.stat(filename).st_mode)
+    return False
 
 
 def set_permissions(filename, permissions):
-    os.chmod(filename, permissions)
+    if os.path.isfile(filename) and (not os.path.isdir(filename)):
+        os.chmod(filename, permissions)
+        return True
+    return False
